:bug: #1723 Move ConcurrentWeakIdentityHashMap out of internal package
